// FRONT-END/src/app/Models/docente/nota.interface.ts
export class Notas  {
    COD_PERSONA: String;
    COD_PERSONA_REPRESENTANTE: String;
    CEDULA: String;
    APELLIDO: String;
    NOMBRE: String;
    DIRECCION: String;
    TELEFONO: String;
    FECHA_NACIMIENTO: Date;
    GENERO: String;
    CORREO: String;
    CORREO_PERSONAL: String;
    COD_PERIODO_LECTIVO: String;
    COD_ALUMNO: 6
    COD_NIVEL_EDUCATIVO: String;
    COD_ASIGNATURA: String;
    COD_PARALELO: String;
    COD_DOCENTE: Number;
    COD_QUIMESTRE: String;
    NOTA1: Number;
    NOTA2: Number;
    NOTA3: Number;
    NOTA4: Number;
    NOTA5: Number;
    NOTA6: Number;
    NOTA7: Number;
    NOTA8: Number;
    NOTA9: Number;
    NOTA10: Number;
    NOTA11: Number;
    NOTA12: Number;
    NOTA13: Number;
    NOTA14: Number;
    NOTA15: Number;
}